refactor(phrases): extract ui-phrases
extract ui phrases
