+++ packages/schemas/package.json
@@ -47,6 +47,7 @@
     "@logto/connector-types": "^0.1.0",
     "@logto/phrases": "^0.1.0",
     "@logto/shared": "^0.1.0",
+    "@logto/ui-phrases": "^0.1.0",
     "zod": "^3.14.3"
   }
 }
