+++ packages/schemas/src/api/error.ts
@@ -1,4 +1,4 @@
-import { LogtoErrorCode } from '@logto/phrases';
+import type { LogtoErrorCode } from '@logto/phrases';
 
 export type RequestErrorMetadata = Record<string, unknown> & {
   code: LogtoErrorCode;
