+++ packages/ui-phrases/README.md
@@ -1,0 +1,11 @@
+# `@logto/ui-phrases`
+
+> TODO: description
+
+## Usage
+
+```
+const uiPhrases = require('@logto/ui-phrases');
+
+// TODO: DEMONSTRATE API
+```
