+++ packages/ui-phrases/src/index.ts
@@ -1,0 +1,16 @@
+import { NormalizeKeyPaths } from '@silverhand/essentials';
+
+import en from './locales/en';
+import zhCN from './locales/zh-cn';
+import { Resource, Language } from './types';
+
+export { Language } from './types';
+
+export type I18nKey = NormalizeKeyPaths<typeof en.translation>;
+
+const resource: Resource = {
+  [Language.English]: en,
+  [Language.Chinese]: zhCN,
+};
+
+export default resource;
