+++ packages/ui-phrases/src/types.ts
@@ -1,0 +1,11 @@
+/* Copied from i18next/index.d.ts */
+export type Resource = Record<Language, ResourceLanguage>;
+
+export type ResourceLanguage = Record<string, ResourceKey>;
+
+export type ResourceKey = string | Record<string, any>;
+
+export enum Language {
+  English = 'en',
+  Chinese = 'zh-CN',
+}
