+++ packages/ui-phrases/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@silverhand/ts-config/tsconfig.base",
+  "compilerOptions": {
+    "outDir": "lib",
+    "declaration": true
+  },
+  "include": ["src"]
+}
