+++ packages/ui/src/components/Button/SocialLinkButton.tsx
@@ -20,7 +20,7 @@ const SocialLinkButton = ({ isDisabled, className, connector, onClick }: Props) 
   const {
     t,
     i18n: { language },
-  } = useTranslation(undefined, { keyPrefix: 'main_flow' });
+  } = useTranslation();
   // TODO: LOG-2393 should fix name[locale] syntax error
   const foundName = Object.entries(name).find(([lang]) => lang === language);
   const localName = foundName ? foundName[1] : name.en;
