+++ packages/ui/src/components/ConfirmModal/AcModal.tsx
@@ -20,7 +20,7 @@ const AcModal = ({
   onConfirm,
   onClose,
 }: ModalProps) => {
-  const { t } = useTranslation(undefined, { keyPrefix: 'main_flow' });
+  const { t } = useTranslation();
 
   return (
     <ReactModal
