+++ packages/ui/src/containers/CreateAccount/index.tsx
@@ -36,7 +36,7 @@ const defaultState: FieldState = {
 };
 
 const CreateAccount = ({ className, autoFocus }: Props) => {
-  const { t } = useTranslation(undefined, { keyPrefix: 'main_flow' });
+  const { t } = useTranslation();
   const { termsValidation } = useTerms();
   const {
     fieldValue,
