+++ packages/ui/src/containers/Passwordless/PhonePasswordless.tsx
@@ -34,7 +34,7 @@ const defaultState: FieldState = { phone: '' };
 const PhonePasswordless = ({ type, autoFocus, className }: Props) => {
   const { setToast } = useContext(PageContext);
   const [showPasswordlessConfirmModal, setShowPasswordlessConfirmModal] = useState(false);
-  const { t } = useTranslation(undefined, { keyPrefix: 'main_flow' });
+  const { t } = useTranslation();
   const { phoneNumber, setPhoneNumber, isValidPhoneNumber } = usePhoneNumber();
   const navigate = useNavigate();
   const { termsValidation } = useTerms();
