+++ packages/ui/src/containers/SocialCreateAccount/index.tsx
@@ -16,7 +16,7 @@ type Props = {
 };
 
 const SocialCreateAccount = ({ connectorId, className }: Props) => {
-  const { t } = useTranslation(undefined, { keyPrefix: 'main_flow' });
+  const { t } = useTranslation();
   const { relatedUser, localSignInMethods, registerWithSocial, bindSocialRelatedUser } =
     useBindSocial();
