+++ packages/ui/src/containers/SocialSignIn/SocialSignInDropdown/index.tsx
@@ -1,4 +1,4 @@
-import { Language } from '@logto/phrases';
+import { Language } from '@logto/ui-phrases';
 import React, { useState, useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
