+++ packages/ui/src/hooks/use-preview.ts
@@ -1,5 +1,5 @@
-import { Language } from '@logto/phrases';
 import { AppearanceMode, ConnectorPlatform } from '@logto/schemas';
+import { Language } from '@logto/ui-phrases';
 import { conditionalString } from '@silverhand/essentials';
 import i18next from 'i18next';
 import { useEffect, useState } from 'react';
