+++ packages/ui/src/i18n/init.ts
@@ -1,5 +1,5 @@
-import resources from '@logto/phrases';
 import { LanguageInfo } from '@logto/schemas';
+import resources from '@logto/ui-phrases';
 import i18next, { InitOptions } from 'i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
 import { initReactI18next } from 'react-i18next';
