+++ packages/ui/src/include.d/react-i18next.d.ts
@@ -2,7 +2,7 @@
 
 // eslint-disable-next-line import/no-unassigned-import
 import 'react-i18next';
-import en from '@logto/phrases/lib/locales/en.js';
+import en from '@logto/ui-phrases/lib/locales/en.js';
 
 declare module 'react-i18next' {
   interface CustomTypeOptions {
